% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{alt_text}
\alias{alt_text}
\title{Generate alt text from the title, subtitle and caption of a ggplot}
\usage{
alt_text(obj)
}
\arguments{
\item{obj}{a ggplot}
}
\description{
Generate alt text from the title, subtitle and caption of a ggplot
}
