% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels.R
\name{gen_alttext}
\alias{gen_alttext}
\title{Generate alt text from the title, subtitle and caption of a ggplot}
\usage{
gen_alttext(obj)
}
\arguments{
\item{obj}{a ggplot}
}
\value{

}
\description{
Generate alt text from the title, subtitle and caption of a ggplot
}
